+++ fhirpy/base/__init__.py
@@ -1,5 +1,6 @@
 from .lib_async import AsyncClient, AsyncReference, AsyncResource, AsyncSearchSet
 from .lib_sync import SyncClient, SyncReference, SyncResource, SyncSearchSet
+from .resource import BaseReference, BaseResource
 from .resource_protocol import ResourceProtocol
 
 __all__ = [
@@ -12,4 +13,6 @@ __all__ = [
     "SyncReference",
     "AsyncReference",
     "ResourceProtocol",
+    "BaseReference",
+    "BaseResource",
 ]